requirements.txt
google-genai
youtube-transcript-api
